students/lab03/carla3791/projects/bookseeker/backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pydantic
python-dotenv
behave
requests
